fix:organização de pastas+bug de login
