--- src/Pages/Chakra_Login/index.jsx
@@ -1,20 +1,0 @@
-import { Box } from "@chakra-ui/react";
-import background from "../../Assets/Images/background.png";
-import ContainerLogin from "../../Components/LandingComponents/Container_Login/index";
-
-const ChakraLoginPage = ({authenticaded,setAuthenticaded}) => {
-  return (
-    <Box
-      w="100vw"
-      h="100vh"
-      bgImg={background}
-      bgPos="center"
-      bgRepeat="no-repeat"
-      bgSize="cover"
-    >
-      <ContainerLogin authenticaded={authenticaded} setAuthenticaded={setAuthenticaded}/>
-    </Box>
-  );
-};
-
-export default ChakraLoginPage;
