--- src/Pages/Chakra_Signup/index.jsx
@@ -1,20 +1,0 @@
-import { Box } from "@chakra-ui/react";
-import background from "../../Assets/Images/background.png";
-import ContainerSignup from "../../Components/LandingComponents/Container_Signup/index";
-
-const ChakraSignupPage = () => {
-  return (
-    <Box
-      w="100vw"
-      h="100vh"
-      bgImg={background}
-      bgPos="center"
-      bgRepeat="no-repeat"
-      bgSize="cover"
-    >
-      <ContainerSignup/>
-    </Box>
-  );
-};
-
-export default ChakraSignupPage;
